Fix: Separate workspaces in multiple directories
