+++ src/AsmResolver.Workspaces.DotNet/Analyzers/Reference/MemberReferenceAnalyser.cs
@@ -1,26 +1,26 @@
 using AsmResolver.DotNet;
 
-namespace AsmResolver.Workspaces.DotNet.Analyzers
+namespace AsmResolver.Workspaces.DotNet.Analyzers.Reference
 {
     /// <summary>
-    /// Analyzes a <see cref="TypeReference"/> for its definitions
+    /// Analyzes a <see cref="MemberReference"/> for its definitions
     /// </summary>
-    public class TypeReferenceAnalyser : ObjectAnalyzer<TypeReference>
+    public class MemberReferenceAnalyser : ObjectAnalyzer<MemberReference>
     {
         /// <inheritdoc />
-        public override void Analyze(AnalysisContext context, TypeReference subject)
+        public override void Analyze(AnalysisContext context, MemberReference subject)
         {
             if(context.Workspace is not DotNetWorkspace workspace)
                 return;
-            
+
             var definition = subject.Resolve();
             if(definition is null || !workspace.Assemblies.Contains(definition.Module.Assembly))
-                return;
-            
+                return; //TODO: Maybe add some warning log?
+
             var index = context.Workspace.Index;
             var node = index.GetOrCreateNode(subject);
             var candidateNode = index.GetOrCreateNode(definition);
-            node.AddRelation(DotNetRelations.ReferenceType, candidateNode);
+            node.AddRelation(DotNetRelations.ReferenceMember, candidateNode);
         }
     }
 }
