+++ src/AsmResolver.Workspaces.DotNet/Analyzers/Reference/TypeReferenceAnalyser.cs
@@ -1,26 +1,26 @@
 using AsmResolver.DotNet;
 
-namespace AsmResolver.Workspaces.DotNet.Analyzers
+namespace AsmResolver.Workspaces.DotNet.Analyzers.Reference
 {
     /// <summary>
-    /// Analyzes a <see cref="MemberReference"/> for its definitions
+    /// Analyzes a <see cref="TypeReference"/> for its definitions
     /// </summary>
-    public class MemberReferenceAnalyser : ObjectAnalyzer<MemberReference>
+    public class TypeReferenceAnalyser : ObjectAnalyzer<TypeReference>
     {
         /// <inheritdoc />
-        public override void Analyze(AnalysisContext context, MemberReference subject)
+        public override void Analyze(AnalysisContext context, TypeReference subject)
         {
             if(context.Workspace is not DotNetWorkspace workspace)
                 return;
-            
+
             var definition = subject.Resolve();
             if(definition is null || !workspace.Assemblies.Contains(definition.Module.Assembly))
-                return; //TODO: Maybe add some warning log?
-            
+                return;
+
             var index = context.Workspace.Index;
             var node = index.GetOrCreateNode(subject);
             var candidateNode = index.GetOrCreateNode(definition);
-            node.AddRelation(DotNetRelations.ReferenceMember, candidateNode);
+            node.AddRelation(DotNetRelations.ReferenceType, candidateNode);
         }
     }
 }
