+++ src/AsmResolver.Workspaces.DotNet/Analyzers/Signature/MethodSignatureBaseAnalyser.cs
@@ -1,8 +1,7 @@
-using AsmResolver.DotNet;
 using AsmResolver.DotNet.Signatures;
 using AsmResolver.DotNet.Signatures.Types;
 
-namespace AsmResolver.Workspaces.DotNet.Analyzers
+namespace AsmResolver.Workspaces.DotNet.Analyzers.Signature
 {
     /// <summary>
     /// Provides a default implementation for an <see cref="MethodSignatureBase"/> analyzer.
