+++ src/AsmResolver.Workspaces.DotNet/Analyzers/Signature/TypeSignatureAnalyser.cs
@@ -1,7 +1,7 @@
 using AsmResolver.DotNet;
 using AsmResolver.DotNet.Signatures.Types;
 
-namespace AsmResolver.Workspaces.DotNet.Analyzers
+namespace AsmResolver.Workspaces.DotNet.Analyzers.Signature
 {
     /// <summary>
     /// Provides a default implementation for an <see cref="TypeSignature"/> analyzer.
